pyproject.toml
[project]
name = "rag"
version = "0.6.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy==2.2.1",
    "pandas==2.2.3",
    "pymupdf==1.25.1",
    "tqdm==4.67.1",
    "torch==2.5.1",
    "marker-pdf==1.2.3",
    "sentence-transformers==3.3.1",
    "giskard[llm]==2.5.0",
    "gguf==0.14.0",
    "accelerate==1.3.0",
    "model2vec==0.3.8",
    "pytest==8.3.4",
    "flagembedding==1.2.11",
    "peft==0.14.0",
    "types-tqdm==4.67.0.20241221",
    "pydantic==2.10.5",
    "langchain==0.3.19",
    "llama-cpp-python==0.3.7",
    "wikipedia-api==0.8.1",
    "tiktoken==0.9.0",
    "gradio==5.20.1",
]

[project.scripts]
main = "rag:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
